requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langchain-google-genai==1.0.6
google-generativeai==0.5.4
python-dotenv==1.0.1
jinja2==3.1.4
requests==2.32.3
beautifulsoup4==4.12.3